\name{ccl3l1data}
\docType{data}
\alias{ccl3l1data}
\title{Data of CCL3L1 gene (The 1000 Genomes Project)}
\description{
This data set includes: segmentation results, population information and CN.
}
\usage{
data(ccl3l1data)

}
\format{
This is a data frame including four columns: \emph{Name} (names of samples), \emph{Pop} (names of populations), \emph{SS} (segmentation scores of samples at the gene) 
and \emph{CN} (CCL3L1 CN obtained by using a Bayesian clustering approach with European-ancestry as prior information).
}
\references{
\url{www.1000genomes.org}
}

\keyword{datasets}

