\name{CNVrd2-package}
\alias{CNVrd2-package}
\docType{package}
\title{
CNVrd2
}
\description{
CNVrd2 uses next-generation sequencing data to measure human gene copy number for multiple samples and indentify SNPs/INDELs which are in linkage disequilibrium with the gene copy number variation.

}
\details{
\tabular{ll}{
Package: \tab CNVrd2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-26\cr
License: \tab GPL-2\cr
Depends: \tab methods\cr
}

}
\author{
Maintainer: Hoang Tan Nguyen <hoangtannguyenvn@gmail.com>

}

\keyword{ package }
