\name{fcgr3bMXL}
\docType{data}
\alias{fcgr3bMXL}
\alias{allGroups}
\alias{bamFiles}
\alias{copynumberGroups}
\alias{objectCNVrd2}
\alias{readCountMatrix}
\alias{resultSegment}
\alias{.Random.seed}
\title{MXL population data (The 1000 Genomes Project)}
\description{
This data set includes: segmentation results \dQuote{resultSegment},
information copy number \dQuote{copynumberGroups} of \emph{FCGR3B} gene.
}
\usage{
data(fcgr3bMXL)

}
\format{All results analysed by the CNVrd2 package.}
\references{
\url{www.1000genomes.org}
}
\keyword{datasets}

