\name{identifyPolymorphicRegion-methods}
\docType{methods}
\alias{identifyPolymorphicRegion-methods}
\alias{identifyPolymorphicRegion,CNVrd2-method}
\title{ Methods for Function \code{identifyPolymorphicRegion}  }
\description{
 Methods for function \code{identifyPolymorphicRegion}  
}
\section{Methods}{
\describe{

\item{\code{signature(Object = "CNVrd2")}}{
}
}}
\keyword{methods}

