\name{numericOrNULL-class}
\Rdversion{1.1}
\docType{class}
\alias{numericOrNULL-class}

\title{Class \code{"numericOrNULL"}}
\description{
 Auxiliary classes; may contain either a numeric vector or NULL [or a call / data.frame or NULL, respectively].
}


\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "numericOrNULL" in the signature.
}

\author{
Hoang Tan Nguyen, Tony R Merriman and MA Black. \email{hoangtannguyenvn@gmail.com}
}

\keyword{classes}
