\name{plotPolymorphicRegion-methods}
\docType{methods}
\alias{plotPolymorphicRegion-methods}
\alias{plotPolymorphicRegion,CNVrd2-method}
\title{Methods for Function \code{plotPolymorphicRegion}  }
\description{
  Methods for function \code{plotPolymorphicRegion}  
}
\section{Methods}{
\describe{

\item{\code{signature(Object = "CNVrd2")}}{

}
}}
\keyword{methods}
