\name{segmentSamples-methods}
\docType{methods}
\alias{segmentSamples-methods}
\alias{segmentSamples,CNVrd2-method}
\title{Method \code{segmentSamples}}
\description{
 Method \code{segmentSamples}  
}
\section{Methods}{
\describe{

\item{\code{signature(Object = "CNVrd2")}}{

}
}}
\keyword{methods}
