\name{segmentSamplesUsingPopInformation-methods}
\docType{methods}
\alias{segmentSamplesUsingPopInformation-methods}
\alias{segmentSamplesUsingPopInformation,CNVrd2-method}
\title{Method \code{segmentSamplesUsingPopInformation}}
\description{
 Method \code{segmentSamplesUsingPopInformation}  
}
\section{Methods}{
\describe{

\item{\code{signature(Object = "CNVrd2")}}{

}
}}
\keyword{methods}
