\name{vectorORfactor-class}
\Rdversion{1.1}
\docType{class}
\alias{vectorORfactor-class}
\title{Class \code{"vectorORfactor"}}
\description{
Auxiliary class
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "vectorORfactor" in the signature.
}

\author{
Hoang Tan Nguyen, Tony Merriman and Mik Black. \email{hoangtannguyenvn@gmail.com}
}

\keyword{classes}
